NuttX: update PWM block with additional -1 last channel number
This change takes an advantage of NuttX PWM drivers prematurely leaving
configuration loops when number of the channel is -1. This number indicates
that all upcoming channels are not used so the driver does not need to
setup them.

The -1 value can be added only if number of used channel is less than the
number of configured channels, otherwise there would be a buffer overflow.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>