resources: update xblk files with type of the parameter
This change updates xblk file with the description of paraeter type. The
new format is Name:value:type when type is optional and not necessary.
This change allows the system to identify real/int parameters in early
build stage and save their names.

The change also presents xblk files in more readable form.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>