RCPgen: link Inputs and Outputs with Nodes statically
This commit removes assignment of Nodes to Inputs and Outputs in
initialization function name_init() and asignes them statically during
variable declaration and definition. The advantage is more clear code.

This change does not affect any other part of the code. Access of blocks'
inputs and outputs remains the same.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>