CodeGen/nuttx: optional ROMFS content can be mounted into /etc on target
EMBEDROMFS NuttX Makefile template option allows to embedded
directory content into target application image in ROMFS.
The content is mounted to /etc in the running NuttX system.

The content of the init.d/rcS file is executed by NSH
at startup. It can be used to setup system or even
load FPGA bitstream from other file stored under /etc,
for example. The start of the pysimCoder model even
with parameters is possible as the last action as well.
It can be run even in background if "main &" is specified.

Signed-off-by: Pavel Pisa <pisa@cmp.felk.cvut.cz>