toolbox/supsisim: refactor SHV part of the code
This refactors SHV part of the code. The SHV tree generation and
client connection is moved to a separate subdirectory and unified
with Class access. The goal is to provide more clear access to SHV
package and simplify the future enhancement and maintenance of the code.

The functionality of SHV generation/connection is not changed.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>