treewide: rework blocks' parameters
The parameters now have their own class PySimParam that stores the
name, value, type and other properties. This change is a step forward
towards generic parameters in block representation. It will be used to
allow many parameters in C representation and also SHV communication
(lists, boolean, integers and so on).

Signed-off-by: Michal Lenc <michallenc@seznam.cz>
Co-authored-by: Stepan Pressl <pressl.stepan@gmail.com>